# openapi.yaml
openapi: 3.1.0
info:
  title: OssPilot Ops API
  version: 0.1.0
paths:
...
  /api/tenant-users/{user_id}/buckets:
    get:
      summary: List buckets granted to a tenant account
      responses:
        "200":
          description: items and total
    put:
      summary: Replace grants and project to tenant API
      responses:
        "200":
          description: items and total
        "409":
          description: bucket limit exceeded
  /api/tenant-users/{user_id}/buckets/{bucket_id}:
    delete:
      summary: Revoke one grant
      responses:
        "204":
          description: revoked
        "404":
          description: grant not found
...
  /api/ops/rgw/restart:
    post:
      summary: Restart RGW daemons
      responses:
        "200":
          description: restart result
        "503":
          description: management API unavailable
  /api/ops/rgw/rolling-restart:
    post:
      summary: Rolling restart RGW daemons
      responses:
        "200":
          description: restart result
        "503":
          description: management API unavailable
...
  /api/tenant-api-access:
    get:
      summary: List tenant Open API access requests
      responses:
        "200":
          description: items and total
        "503":
          description: tenant projection is not configured
  /api/tenant-api-access/{account_id}:
    get:
      summary: Get Open API access for an ops tenant account
      responses:
        "200":
          description: access record
        "404":
          description: account or access record not found
  /api/tenant-api-access/{account_id}/approve:
    post:
      summary: Approve a pending Open API access request
      responses:
        "200":
          description: approved
        "409":
          description: not pending
  /api/tenant-api-access/{account_id}/reject:
    post:
      summary: Reject a pending Open API access request
      responses:
        "200":
          description: rejected
        "409":
          description: not pending
  /api/tenant-api-access/{account_id}/disable:
    post:
      summary: Disable approved Open API access
      responses:
        "200":
          description: disabled
        "409":
          description: not approved
...